Change location of the useBindPatientToDoctor function for now
