+++ NaZdrowie/src/properties/styles/index.ts
@@ -1,8 +1,8 @@
+export * from "./authenticationScreenStyle";
 export * from "./buttonStyle";
 export * from "./cardStyle";
-export * from "./authenticationScreenStyle";
-export * from "./mainStyle";
 export * from "./commentStyle";
 export * from "./generalStyle";
 export * from "./inputStyle";
+export * from "./mainStyle";
 export * from "./overlayStyle";
